BUG: find_date_cols: issue with detection of datetime in columns
find_date_cols:
Static format of y-m-d was used to detect and find if the column is of type datetime, as a result all columns failed to pass correctly, hence resulting in a "NULL" selection list, have updated the format to be more flexible.
This issue only appears to crop up on streamlit 1.2+ as 1.1+ did not have an issue.